requirements.txt
crewai>=0.1.32
langchain>=0.0.300
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pydantic>=1.10.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 